+++ doc/source/changelog.rst
@@ -17,6 +17,7 @@ master: 2.5.0
 * Support explicitly created initializer lists as arguments
 * Pass instances by-ref in Numba traces
 * Support non-POD by-value returns in Numba traces
+* Nullify derived class Python proxy when the C++ object is deleted
 
 
 2022-10-03: 2.4.1